fix (#16692)
(cherry picked from commit bd19973a6a0c5f67cf1884f29cd4c6cb7ef10613)
